+++ app/controllers/api/v1/AttachmentsController.js
@@ -194,6 +194,10 @@ export default class AttachmentsController {
         throw new ValidationException('Invalid format value');
       }
 
+      if ('variant' in query && query.variant !== 'hdr') {
+        throw new ValidationException('Invalid variant value');
+      }
+
       const width = 'width' in query ? Number.parseInt(query.width, 10) : undefined;
       const height = 'height' in query ? Number.parseInt(query.height, 10) : undefined;
 
@@ -206,6 +210,7 @@ export default class AttachmentsController {
 
       const asRedirect = 'redirect' in query;
       const withDownload = 'download' in query;
+      const preferHdr = query.variant === 'hdr';
 
       const attachment = await dbAdapter.getAttachmentById(attId);
 
@@ -235,7 +240,8 @@ export default class AttachmentsController {
       } else {
         // Visual types, 'image' and 'video'
 
-        const previews = attachment.previews[type];
+        const useHdrPreview = type === 'image' && preferHdr && attachment.previews.imageHDR;
+        const previews = useHdrPreview ? attachment.previews.imageHDR : attachment.previews[type];
         const {
           variant,
           width: resWidth,
@@ -248,11 +254,17 @@ export default class AttachmentsController {
         response.width = prv.w;
         response.height = prv.h;
 
+        if (useHdrPreview) {
+          response.variant = 'hdr';
+        }
+
         // With imgproxy, we can resize images (except some types) and change
         // their format
         if (
           useImgProxy &&
           type === 'image' &&
+          // imgproxy doesn't support HDR images
+          !useHdrPreview &&
           // We don't need to resize SVGs
           attachment.fileExtension !== 'svg' &&
           // We should not resize (probably) animated legacy GIFs